+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Re-worked the dialog showing waveform overlay at looppoints to be independent (modeless). (TODO)
 - Batch process naming "Kill" loops/cues to "Remove" instead.
+- Loop overlay drawing on separate wxPanel.
 
 ### Fixed